refactor: massive refactoring to make the codebase more readable and future proof.
