+++ src/main/java/com/polytech/webscraipper/repositories/DocumentRepository.java
@@ -1,0 +1,8 @@
+package com.polytech.webscraipper.repositories;
+
+import com.polytech.webscraipper.dto.DocumentDto;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface DocumentRepository extends MongoRepository<DocumentDto,String>{
+
+}
